requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
pypdf
docx2txt
streamlit
faiss-cpu